pyproject.toml
[project]
name = "vnpy_polygon"
dynamic = ["version"]
description = "Polygon.io datafeed for VeighNa quant trading framework."
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Xiaoyou Chen", email = "xiaoyou.chen@mail.vnpy.com"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Programming Language :: Python :: Implementation :: CPython",
    "Natural Language :: Chinese (Simplified)",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "polygon-api-client"
]
keywords = ["quant", "quantitative", "investment", "trading", "algotrading"]

[project.urls]
"Homepage" = "https://www.vnpy.com"
"Documentation" = "https://www.vnpy.com/docs"
"Changes" = "https://github.com/vnpy/vnpy_polygon/blob/master/CHANGELOG.md"
"Source" = "https://github.com/vnpy/vnpy_polygon/"
"Forum" = "https://www.vnpy.com/forum"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "vnpy_polygon/__init__.py"
pattern = "__version__ = ['\"](?P<version>[^'\"]+)['\"]"

[tool.hatch.build.targets.wheel]
packages = ["vnpy_polygon"]
include-package-data = true

[tool.hatch.build.targets.sdist]
include = ["vnpy_polygon*"] 

[tool.ruff]
target-version = "py310"
output-format = "full"

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "E",  # pycodestyle error
    "F",  # pyflakes
    "UP",  # pyupgrade
    "W",  # pycodestyle warning
]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true